+++ src/tightdb/group_shared.hpp
@@ -23,6 +23,8 @@
 #include <limits>
 
 #include <tightdb/util/features.h>
+#include <tightdb/util/thread.hpp>
+#include <tightdb/util/platform_specific_condvar.hpp>
 #include <tightdb/group.hpp>
 //#include <tightdb/commit_log.hpp>
 
@@ -346,6 +348,7 @@ public:
 
 private:
     struct SharedInfo;
+    struct ReadCount;
     struct ReadLockInfo {
         uint_fast64_t   m_version;
         uint_fast32_t   m_reader_idx;
@@ -372,7 +375,12 @@ private:
         transact_Writing
     };
     TransactStage m_transact_stage;
-    struct ReadCount;
+#ifndef _WIN32
+    util::PlatformSpecificCondVar m_room_to_write;
+    util::PlatformSpecificCondVar m_work_to_do;
+    util::PlatformSpecificCondVar m_daemon_becomes_ready;
+    util::PlatformSpecificCondVar m_new_commit_available;
+#endif
 
     // Ring buffer managment
     bool        ringbuf_is_empty() const TIGHTDB_NOEXCEPT;